范本关联表修改 保证金bug

+++ blade-service-api/blade-contract-api/src/main/java/org/springblade/contract/entity/ContractBondEntity.java
@@ -61,7 +61,7 @@ public class ContractBondEntity extends BaseEntity {
 	 * 关联合同数量
 	 */
 	@ApiModelProperty(value = "关联合同数量")
-	private BigDecimal contractNumber;
+	private Integer contractNumber;
 	/**
 	 * 币种
 	 */
