+++ blade-service-api/blade-contract-api/src/main/java/org/springblade/contract/entity/ContractBondPlanEntity.java
@@ -56,7 +56,7 @@ public class ContractBondPlanEntity extends BaseEntity {
 	 * 关联合同数量
 	 */
 	@ApiModelProperty(value = "关联合同数量")
-	private BigDecimal contractNumber;
+	private Integer contractNumber;
 	/**
 	 * 币种
 	 */
